LayoutTests/fast/viewport/viewport-9.html

<head>
    <title>Viewport META with "maximum-scale=0, initial-scale=3".</title>
    <meta name="viewport" content="maximum-scale=0, initial-scale=3">
    <script>
        function test() {
            if (window.testRunner) {
                testRunner.dumpAsText();
                alert(internals.configurationForViewport(document, 1, 320, 480, 320, 352));
            }
        }
    </script>
</head>
<body onload="test();">
