Safari-611.1.10/LayoutTests/fast/viewport/viewport-9-expected.txt

ALERT: viewport size 1280x1408 scale 0.25 with limits [0.25, 0.25] and userScalable true
